Change Constructors & Getters-Setters to Annotations @Data, @AllArgsConstructor & @NoArgsConstructor
