+++ src/main/java/com/example/App/DTO/DepartamentoDTO.java
@@ -5,7 +5,7 @@
 package com.example.App.DTO;
 
 
-import com.example.App.clases.*;
+import com.example.App.entities.*;
 import java.util.List;
