+++ src/main/java/com/example/App/Repository/ClienteRepository.java
@@ -1,12 +1,7 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
- */
 package com.example.App.Repository;
 
-import com.example.App.clases.Cliente;
+import com.example.App.entities.Cliente;
 import org.springframework.data.jpa.repository.JpaRepository;
-
 
 public interface ClienteRepository extends JpaRepository<Cliente, Long> {
