--- src/main/java/com/example/App/clases/Cotizacion.java
@@ -1,56 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package com.example.App.clases;
-
-
-import java.io.Serializable;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import org.springframework.web.multipart.MultipartFile;
-
-@Entity
-@Table
-
-public class Cotizacion {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_cotizacion;
-    private String titulo ;
-   // private MultipartFile pdf;
-
-    public Cotizacion() {
-    }
-
- 
-
-    public String getTitulo() {
-        return titulo;
-    }
-
-    public void setTitulo(String titulo) {
-        this.titulo = titulo;
-    }
-/*  public MultipartFile getPdf() {
-        return pdf;
-    }
-
-    public void setPdf(MultipartFile pdf) {
-        this.pdf = pdf;
-    }
-*/
-
-    public int getId_cotizacion() {
-        return id_cotizacion;
-    }
-
-    public void setId_cotizacion(int id_cotizacion) {
-        this.id_cotizacion = id_cotizacion;
-    }
-    
-    
-}
