+++ src/main/java/com/example/App/entities/Cotizacion.java
@@ -1,0 +1,21 @@
+package com.example.App.entities;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+public class Cotizacion {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id_cotizacion;
+    private String titulo;
+    // private MultipartFile pdf;
+}
