+++ src/main/java/com/example/App/entities/Rol_Usuario.java
@@ -1,0 +1,24 @@
+package com.example.App.entities;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.OneToMany;
+
+@Entity
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Rol_Usuario {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id_rol_usuario;
+    private String rol;
+    private String descripcion;
+}
